Rework infection command and split it up in four parts
