+++ tests/Process/Runner/InitialTestDidNotPassTest.php
@@ -35,51 +35,38 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Process\Runner;
 
-use Infection\EventDispatcher\EventDispatcherInterface;
-use Infection\Events\InitialTestCaseCompleted;
-use Infection\Events\InitialTestSuiteFinished;
-use Infection\Events\InitialTestSuiteStarted;
-use Infection\Process\Builder\InitialProcessBuilder;
-use Infection\Process\Runner\InitialTestsRunner;
-use PHPUnit\Framework\MockObject\MockObject;
+use Infection\Process\Runner\InitialTestDidNotPass;
+use Infection\TestFramework\AbstractTestFrameworkAdapter;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Process\Process;
 
-final class InitialTestsRunnerTest extends TestCase
+final class InitialTestDidNotPassTest extends TestCase
 {
-    public function test_it_dispatches_events(): void
+    public function test_log_initial_tests_do_not_pass(): void
     {
-        /** @var MockObject|Process $process */
         $process = $this->createMock(Process::class);
+        $process->expects($this->once())->method('getExitCode')->willReturn(0);
+        $process->expects($this->once())->method('getOutput')->willReturn('output string');
+        $process->expects($this->once())->method('getErrorOutput')->willReturn('error string');
+        $process->expects($this->once())->method('getCommandLine')->willReturn('vendor/bin/phpunit --order=random');
 
-        $process->expects($this->once())
-            ->method('run')
-            ->with($this->callback(static function ($processCallback): bool {
-                $processCallback(Process::OUT);
+        $testFrameworkAdapter = $this->createMock(AbstractTestFrameworkAdapter::class);
+        $testFrameworkAdapter->expects($this->once())->method('getName')->willReturn('phpunit');
+        $testFrameworkAdapter->expects($this->once())->method('getInitialTestsFailRecommendations')->willReturn('-');
 
-                return true;
-            }));
-        $process->expects($this->once())
-            ->method('getOutput')
-            ->willReturn('foo');
+        $error = implode("\n", [
+            'Project tests must be in a passing state before running Infection.',
+            '-',
+            'phpunit reported an exit code of 0.',
+            'Refer to the phpunit\'s output below:',
+            'STDOUT:',
+            'output string',
+            'STDERR:',
+            'error string',
+        ]);
 
-        $processBuilder = $this->createMock(InitialProcessBuilder::class);
-        $processBuilder->method('getProcessForInitialTestRun')
-            ->with('', false, [])
-            ->willReturn($process);
+        $exception = InitialTestDidNotPass::fromProcessAndAdapter($process, $testFrameworkAdapter);
 
-        /** @var MockObject|EventDispatcherInterface $eventDispatcher */
-        $eventDispatcher = $this->createMock(EventDispatcherInterface::class);
-        $eventDispatcher->expects($this->exactly(3))
-            ->method('dispatch')
-            ->withConsecutive(
-                $this->isInstanceOf(InitialTestSuiteStarted::class),
-                $this->isInstanceOf(InitialTestCaseCompleted::class),
-                $this->isInstanceOf(InitialTestSuiteFinished::class)
-            );
-
-        $testRunner = new InitialTestsRunner($processBuilder, $eventDispatcher);
-
-        $testRunner->run('', false);
+        $this->assertSame($error, $exception->getMessage());
     }
 }
